+++ src/main/java/fll/web/playoff/Playoff.java
@@ -175,11 +175,17 @@ public final class Playoff {
                                           final WinnerType winnerCriteria,
                                           final Team teamA,
                                           final Team teamB,
-                                          final TeamScore teamBScore,
+                                          final PerformanceTeamScore teamBScore,
                                           final int runNumber)
       throws SQLException {
-    final DatabaseTeamScore teamAScore = new DatabaseTeamScore(tournament, teamA.getTeamNumber(), runNumber,
-                                                               connection);
+    final @Nullable PerformanceTeamScore teamAScore = DatabasePerformanceTeamScore.fetchTeamScore(tournament,
+                                                                                                  teamA.getTeamNumber(),
+                                                                                                  runNumber,
+                                                                                                  connection);
+    if (null == teamAScore) {
+      return null;
+    }
+
     final Team retval = pickWinner(performanceElement, tiebreakerElement, winnerCriteria, teamA, teamAScore, teamB,
                                    teamBScore);
     return retval;
@@ -194,9 +200,9 @@ public final class Playoff {
                                            final List<TiebreakerTest> tiebreakerElement,
                                            final WinnerType winnerCriteria,
                                            final Team teamA,
-                                           final TeamScore teamAScore,
+                                           final PerformanceTeamScore teamAScore,
                                            final Team teamB,
-                                           final TeamScore teamBScore) {
+                                           final PerformanceTeamScore teamBScore) {
 
     // teamA can actually be a bye here in the degenerate case of a 3-team
     // tournament with 3rd/4th place brackets enabled...
@@ -206,29 +212,24 @@ public final class Playoff {
         || Team.TIE.equals(teamB)) {
       return null;
     } else {
-      if (teamAScore.scoreExists()
-          && teamBScore.scoreExists()) {
-        final boolean noshowA = teamAScore.isNoShow();
-        final boolean noshowB = teamBScore.isNoShow();
-        if (noshowA
-            && !noshowB) {
-          return teamB;
-        } else if (!noshowA
-            && noshowB) {
-          return teamA;
-        } else {
-          final double scoreA = perf.evaluate(teamAScore);
-          final double scoreB = perf.evaluate(teamBScore);
-          if (FP.lessThan(scoreA, scoreB, TIEBREAKER_TOLERANCE)) {
-            return WinnerType.HIGH == winnerCriteria ? teamB : teamA;
-          } else if (FP.lessThan(scoreB, scoreA, TIEBREAKER_TOLERANCE)) {
-            return WinnerType.HIGH == winnerCriteria ? teamA : teamB;
-          } else {
-            return evaluateTiebreaker(tiebreakerElement, teamA, teamAScore, teamB, teamBScore);
-          }
-        }
+      final boolean noshowA = teamAScore.isNoShow();
+      final boolean noshowB = teamBScore.isNoShow();
+      if (noshowA
+          && !noshowB) {
+        return teamB;
+      } else if (!noshowA
+          && noshowB) {
+        return teamA;
       } else {
-        return null;
+        final double scoreA = perf.evaluate(teamAScore);
+        final double scoreB = perf.evaluate(teamBScore);
+        if (FP.lessThan(scoreA, scoreB, TIEBREAKER_TOLERANCE)) {
+          return WinnerType.HIGH == winnerCriteria ? teamB : teamA;
+        } else if (FP.lessThan(scoreB, scoreA, TIEBREAKER_TOLERANCE)) {
+          return WinnerType.HIGH == winnerCriteria ? teamA : teamB;
+        } else {
+          return evaluateTiebreaker(tiebreakerElement, teamA, teamAScore, teamB, teamBScore);
+        }
       }
     }
   }
@@ -1507,26 +1508,26 @@ public final class Playoff {
           + " found that it's already finished, skipping");
       return;
     } else if (teamAscoreExists) {
-      final TeamScore teamBscore = new DummyTeamScore(teamBteamNumber, performanceRunNumberToEnter, simpleGoals,
-                                                      enumGoals, true, false);
+      final PerformanceTeamScore teamBscore = new DummyTeamScore(teamBteamNumber, performanceRunNumberToEnter,
+                                                                 simpleGoals, enumGoals, true, false);
       Queries.insertPerformanceScore(runMetadataFactory, connection, datasource, description, tournament, true,
                                      teamBscore);
 
     } else if (teamBscoreExists) {
-      final TeamScore teamAscore = new DummyTeamScore(teamAteamNumber, performanceRunNumberToEnter, simpleGoals,
-                                                      enumGoals, true, false);
+      final PerformanceTeamScore teamAscore = new DummyTeamScore(teamAteamNumber, performanceRunNumberToEnter,
+                                                                 simpleGoals, enumGoals, true, false);
       Queries.insertPerformanceScore(runMetadataFactory, connection, datasource, description, tournament, true,
                                      teamAscore);
     } else {
       // initial value score
-      final TeamScore teamAscore = new DummyTeamScore(teamAteamNumber, performanceRunNumberToEnter, simpleGoals,
-                                                      enumGoals, false, false);
+      final PerformanceTeamScore teamAscore = new DummyTeamScore(teamAteamNumber, performanceRunNumberToEnter,
+                                                                 simpleGoals, enumGoals, false, false);
       Queries.insertPerformanceScore(runMetadataFactory, connection, datasource, description, tournament, true,
                                      teamAscore);
 
       // no show
-      final TeamScore teamBscore = new DummyTeamScore(teamBteamNumber, performanceRunNumberToEnter, simpleGoals,
-                                                      enumGoals, true, false);
+      final PerformanceTeamScore teamBscore = new DummyTeamScore(teamBteamNumber, performanceRunNumberToEnter,
+                                                                 simpleGoals, enumGoals, true, false);
       Queries.insertPerformanceScore(runMetadataFactory, connection, datasource, description, tournament, true,
                                      teamBscore);
     }
@@ -1727,7 +1728,7 @@ public final class Playoff {
                                         final List<TiebreakerTest> tiebreakerElement,
                                         final int teamNumber,
                                         final int runNumber,
-                                        final TeamScore teamScore)
+                                        final PerformanceTeamScore teamScore)
       throws SQLException, ParseException {
     if (LOGGER.isTraceEnabled()) {
       LOGGER.trace("Updating playoff score for team: "