+++ src/main/java/fll/web/playoff/TeamScore.java
@@ -5,41 +5,13 @@
  */
 package fll.web.playoff;
 
-import org.checkerframework.checker.initialization.qual.UnknownInitialization;
 import org.checkerframework.checker.nullness.qual.Nullable;
 import org.checkerframework.dataflow.qual.SideEffectFree;
 
 /**
- * Represents a score for a team. Only the values of simple goals are available
- * through this object.
+ * Represents a score for a team.
  */
-public abstract class TeamScore {
-
-  /**
-   * Run number used for team scores that are not performance scores.
-   */
-  public static final int NON_PERFORMANCE_RUN_NUMBER = -1;
-
-  /**
-   * Create a non-performance TeamScore for the specified team.
-   *
-   * @param teamNumber {@link #getTeamNumber()}
-   */
-  public TeamScore(final int teamNumber) {
-    this(teamNumber, NON_PERFORMANCE_RUN_NUMBER);
-  }
-
-  /**
-   * Create a performance TeamScore for the specified team and run number.
-   *
-   * @param teamNumber the team the score is for
-   * @param runNumber 1-based run number
-   */
-  public TeamScore(final int teamNumber,
-                   final int runNumber) {
-    this.teamNumber = teamNumber;
-    this.runNumber = runNumber;
-  }
+public interface TeamScore {
 
   /**
    * The team that this score is for.
@@ -47,78 +19,33 @@ public abstract class TeamScore {
    * @return the team
    */
   @SideEffectFree
-  public final int getTeamNumber(@UnknownInitialization(TeamScore.class) TeamScore this) {
-    return teamNumber;
-  }
-
-  private final int teamNumber;
-
-  /**
-   * Check if the score exists. If it doesn't exist, the other score methods
-   * will throw a RuntimeException
-   *
-   * @return true if the score exists
-   */
-  public abstract boolean scoreExists();
+  int getTeamNumber();
 
   /**
    * Is this score a no show?
    *
    * @return true if this score is a no show
    */
-  public abstract boolean isNoShow();
-
-  /**
-   * Is this score a bye?
-   *
-   * @return true if this score is a no show
-   */
-  public abstract boolean isBye();
-
-  /**
-   * Is the score verified.
-   * 
-   * @return true if this score has been verified
-   */
-  public abstract boolean isVerified();
-
-  /**
-   * When the score is entered from a tablet where the table is known, this has
-   * the table name, otherwise the value is "ALL" meaning that the person entering
-   * the score has "all tables" selected.
-   * 
-   * @return the table that the score was entered from.
-   */
-  public abstract String getTable();
-
-  /**
-   * What run do these scores apply to?
-   * This is a 1-based number. It will be {@link #NON_PERFORMANCE_RUN_NUMBER} if
-   * this is not a performance score.
-   *
-   * @return the run for the scores
-   */
   @SideEffectFree
-  public final int getRunNumber(@UnknownInitialization(TeamScore.class) TeamScore this) {
-    return runNumber;
-  }
-
-  private final int runNumber;
+  boolean isNoShow();
 
   /**
    * The raw score for a particular simple goal, as a double.
    *
    * @param goalName the goal to get the score for
-   * @return the score, NaN if there is no score for the specified name
+   * @return the score, NaN if there is no score for the specified goal name
    */
-  public abstract double getRawScore(String goalName);
+  @SideEffectFree
+  double getRawScore(String goalName);
 
   /**
    * The raw score for a particular enumerated goal, as a String.
    *
    * @param goalName the goal to get the score for
-   * @return the score, null if there is no score for the specified name
+   * @return the score, null if there is no score for the specified goal name
    */
-  public abstract @Nullable String getEnumRawScore(String goalName);
+  @SideEffectFree
+  @Nullable
+  String getEnumRawScore(String goalName);
 
 }